+++ Views/UpdateProgressPage.xaml.cs
@@ -2,7 +2,7 @@ using Library.ViewModels;
 
 namespace Library.Views;
 
-public partial class UpdateProgressPage : ContentPage
+public partial class UpdateProgressPage : BasePage
 {
     public UpdateProgressPage(UpdateProgressViewModel viewModel)
     {